Added resource controllers, ability to only see your own listings, backend security for listing routes, show-listing functionality, page redirects after email verification and login, and beautification messages for when there are no listings.

+++ app/Http/Controllers/HouseController.php
@@ -2,14 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+// use Illuminate\Http\Request;
 use App\Http\Requests\HouseCreateAndEditRequest;
 use App\Models\House;
 
 class HouseController extends Controller
 {
+    public function __construct(){
+        // $this->middleware('auth', ['except' => [
+        //     'allListings', 'show'
+        // ]]);
+        $this->middleware('auth');
+    }
+
     public function index(){
-        $houses = House::all();
+        $houses = auth()->user()->houses()->orderBy('created_at')->get();
         return view('houses.index')->with('houses', $houses);
     }
 
@@ -18,27 +25,12 @@ class HouseController extends Controller
     }
 
     public function store(HouseCreateAndEditRequest $request){
-        //Store the house listing in the database.
-        // dd($request->all());
-
-        // try{
-        //     House::create($request->all());
-        // } catch(\Exception $e){
-        //     return redirect()->back()->with('error', 'Something went wrong.');
-        // }
-
-        // $request->validate([
-        //     'address_street' => 'required|max:200',
-        //     'address_city' => 'required|max:200',
-        //     'address_state' => 'required|max:2',
-        //     'address_zip' => 'required|max:10|regex:/^\d{5}([\-]?\d{4})?$/',
-        //     'price' => 'required|regex:/^\d+(\.\d{1,2})?$/',
-        //     'description' => 'required'
-        // ]);
-
-        House::create($request->all());
-
+        auth()->user()->houses()->create($request->all());
         return redirect()->back()->with('message', 'House listing created.');
+    }
+
+    public function show(House $house){
+        return view('houses.show')->with('house', $house);
     }
 
     public function edit(House $house){
@@ -46,34 +38,34 @@ class HouseController extends Controller
     }
 
     public function update(HouseCreateAndEditRequest $request, House $house){
-        //Update house listing in the database:
-        // dd($request->all());
-
-        $house->update([
-            'address_street' => $request->address_street,
-            'address_city' => $request->address_city,
-            'address_state' => $request->address_state,
-            'address_zip' => $request->address_zip,
-            'price' => $request->price,
-            'description' => $request->description,
-            'sold' => $request->sold
-        ]);
-
-        return redirect(route('my-listings'))->with('message', 'House listing updated.');
+        if(auth()->user()->id === $house->user_id){
+            $house->update([
+                'address_street' => $request->address_street,
+                'address_city' => $request->address_city,
+                'address_state' => $request->address_state,
+                'address_zip' => $request->address_zip,
+                'price' => $request->price,
+                'description' => $request->description,
+                'sold' => $request->sold
+            ]);
+            return redirect(route('houses.index'))->with('message', 'House listing updated.');
+        } else {
+            return redirect()->back()->with('error', 'You may only edit your own listings!');
+        }
     }
 
-    public function details(){
-        return "House details.";
+    public function destroy(House $house){
+        if(auth()->user()->id === $house->user_id){
+            $house->delete();
+            return redirect()->back()->with('message', 'House listing deleted.');
+        } else {
+            return redirect()->back()->with('error', 'You may only delete your own listings!');
+        }
     }
 
-    public function myListings(){
-        $houses = House::all();
-        return view('houses.my-listings')->with('houses', $houses);
-    }
-
-    public function delete(House $house){
-        $house->delete();
-        return redirect()->back()->with('message', 'House listing deleted.');
+    public function allListings(){
+        $houses = House::all()->sortBy('created_at');
+        return view('houses.all-listings')->with('houses', $houses);
     }
 
 }
